requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
playwright
pandas
openpyxl
mysql-connector-python
Authlib
requests
timezonefinder
pytesseract
Pillow
gunicorn
python-dotenv
Werkzeug==3.0.3